CTB Comments, move to_memmap to np_utils
